Refactor Room's page state
This commit completely rewrites the Room module to split the model from
the page itself from the room data that will be shared.

Also this models a state machine by introducing the page's states
(reflected on the `Room.Model`).

This way we can have more control on when and how we deal with certain
events / states.

The pattern I'm following on the `update` function is something like

(CurrentState, Msg)

So it replicate what a state machine would do.

The code still a little bit messy, a clean up will happen after the
first version is up

Also sneaked some layout changes in there just for fun, no harm done