Add playlist feature
Now we store a list of videoID to represent a playlist that will be
played in order (for now) and the current video index to keep track
which video is being played.

Based on the event AddVideo we append the new video ID to the list and
broadcast the message to other clients on the room. For
now, we're just not adding it when the URL for the video is invalid. We
will give a proper message on the field on the future

Once the video has ended, we change the current video index or do
nothing if the index is not found on the current playlist.